backend/ansd_suivi/requirements.txt

djangorestframework
django-cors-headers
# Ajoutez d'autres dépendances si nécessaire au fur et à mesure
# psycopg2-binary  # Si vous utilisez PostgreSQL en production